% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devianceGamma.R
\name{devianceGamma}
\alias{devianceGamma}
\title{Function to calculate deviance for model predictions assuming a Gamma distribution.}
\usage{
devianceGamma(y, y_hat)
}
\arguments{
\item{y}{a numeric vector of observations.}

\item{y_hat}{a numeric vector of predictions for y (must have same length as y.)}
}
\value{
a numeric vector.
}
\description{
This function calculates a deviance measure for model predictions assuming a Gamma distribution.
}
\examples{
# n <- 1000
# 
# true <- runif(n, 100, 2000)
# observed <- rgamma(n, true)
# predicted <- exp(log(1050)/2 + log(true)/2 + rnorm(n, sd=0.3))
# 
# plot(observed, predicted)
# 
# devs <- devianceGamma(observed, predicted)
# sum(devs)
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
