% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devianceNormal.R
\name{devianceNormal}
\alias{devianceNormal}
\title{Function to calculate deviance for model predictions assuming a Normal distribution.}
\usage{
devianceNormal(Y, Y_hat)
}
\arguments{
\item{Y}{a numeric vector of observations.}

\item{Y_hat}{a numeric vector of predictions for Y (must have same length as Y.)}
}
\value{
a numeric vector.
}
\description{
This function calculates a deviance measure for model predictions assuming a Normal distribution.
}
\examples{
# n <- 1000
# 
# true <- rnorm(n)
# observed <- rnorm(n, true, sd=0.1)
# predicted <- true/2 + rnorm(n, sd=0.2)
# 
# plot(observed, predicted)
# 
# devs <- devianceNormal(observed, predicted)
# sum(devs)
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
