% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviancePoisson.R
\name{deviancePoisson}
\alias{deviancePoisson}
\title{Function to calculate deviance for model predictions assuming a Poisson distribution.}
\usage{
deviancePoisson(Y, Y_hat)
}
\arguments{
\item{Y}{a numeric vector of observations.}

\item{Y_hat}{a numeric vector of predictions for Y (must have same length as Y.)}
}
\value{
a numeric vector.
}
\description{
This function calculates a deviance measure for model predictions assuming a Poisson distribution.
}
\examples{
# n <- 1000
# 
# true <- runif(n, 0, 3)
# observed <- rpois(n, true)
# predicted <- exp(log(1.5)/2 + log(true)/2 + rnorm(n, sd=0.1))
# 
# plot(observed, predicted)
# 
# devs <- deviancePoisson(observed, predicted)
# sum(devs)
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
