% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfBinary-data.R
\docType{data}
\name{dfBinary}
\alias{dfBinary}
\title{Random data for binary classification example.}
\format{An object of class \code{data.frame} with 1000 rows and 4 columns.}
\usage{
data("dfBinary")
}
\description{
Random data for binary classification example.
}
\examples{
data("dfBinary")
}
\keyword{datasets}
