% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfGamma-data.R
\docType{data}
\name{dfGamma}
\alias{dfGamma}
\title{Random data for Gamma distribution regression example.}
\format{A data.frame with 1,000 rows and 4 variables:
\itemize{
  \item x1: feature 1
  \item x2: feature 2
  \item x3: feature 3
  \item target: Gamma distributed target variable
}}
\usage{
data("dfGamma")
}
\description{
Random data for Gamma distribution regression example.
}
\examples{
data("dfGamma")
}
\keyword{datasets}
