% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfPoisson-data.R
\docType{data}
\name{dfPoisson}
\alias{dfPoisson}
\title{Random data for Poisson process regression example.}
\format{A data.frame with 1,000 rows and 4 variables:
\itemize{
  \item x1: feature 1
  \item x2: feature 2
  \item x3: feature 3
  \item exposure: length of time of Poisson process
  \item target: Poisson process target variable
}}
\usage{
data("dfPoisson")
}
\description{
Random data for Poisson process regression example.
}
\examples{
data("dfPoisson")
}
\keyword{datasets}
