% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfTweedie-data.R
\docType{data}
\name{dfTweedie}
\alias{dfTweedie}
\title{Random data for compound Poisson-Gamma process regression example.}
\format{A data.frame with 1,000 rows and 5 variables:
\itemize{
  \item x1: feature 1
  \item x2: feature 2
  \item x3: feature 3
  \item exposure: length of time of Poisson part of the compound Poisson-Gamma process
  \item target: target variable
}}
\usage{
data("dfTweedie")
}
\description{
Random data for compound Poisson-Gamma process regression example.
}
\examples{
data("dfTweedie")
}
\keyword{datasets}
