% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise a vector}
\usage{
normalise(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Normalise a vector by subtracting mean and dividing by standard deviation
}
\examples{
normalise(0:100)
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
