% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehotEncode.R
\name{onehotEncode}
\alias{onehotEncode}
\title{One-hot encode a factor}
\usage{
onehotEncode(x)
}
\arguments{
\item{x}{A factor}
}
\value{
A matrix of ones and zeros with length(x) rows and length(levels(x)) columns. If there is only one level returns a vector.
}
\description{
One-hot encode a factor
}
\examples{
normalise(0:100)
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
