% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFamily.R
\name{outputFamily}
\alias{outputFamily}
\title{Function to calculate deviance for model predictions assuming a Bernoulli distribution.}
\usage{
devianceBernoulli(y, y_hat)
}
\arguments{
\item{y}{a numeric vector of observations (0s and 1s.)}

\item{y_hat}{a numeric vector of predictions (between 0 and 1) for y (must have same length as y.)}
}
\value{
a numeric vector.
}
\description{
This function calculates a deviance measure for model predictions assuming a Bernoulli distribution.
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
