% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.EdNetModel.R
\name{predict.EdNetModel}
\alias{predict.EdNetModel}
\title{Predict for EdNetModel objects}
\usage{
predict.EdNetModel(object, newdata=NULL)
}
\arguments{
\item{object}{A matrix with rows as training examples and columns as input features}

\item{newdata}{A matrix with rows as training examples and columns as target values}
}
\value{
A vector or matrix of predictions.
}
\description{
Predict for EdNetModel objects
}
\examples{
# No example yet
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
