% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictedClass.R
\name{predictedClass}
\alias{predictedClass}
\title{Converts class probabilities into a predicted class}
\usage{
predictedClass(classProbs, classes=NULL)
}
\arguments{
\item{classProbs}{A matrix of probabilities with rows representing train/dev/test examples and columns representing classes.}

\item{classes}{An optional character vector representing the names of the classes. If left blank the names are taken from the column names of classProbs.}
}
\value{
A factor.
}
\description{
For multi-class learning, takes matrix of class probabilities and returns the predicted (most probable) class.
}
\examples{
# No example yet
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
