% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relu.R
\name{relu}
\alias{relu}
\title{Compute relu function on vector}
\usage{
relu(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector the same length as x.
}
\description{
Compute relu function on vector
}
\examples{
# No example yet
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
