% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Compute sigmoid function on vector}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector the same length as x.
}
\description{
Compute sigmoid function on vector
}
\examples{
# No example yet
}
\author{
Edwin Graham <edwingraham1984@gmail.com>
}
