% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfBinary-data.R
\docType{data}
\name{dfBinary}
\alias{dfBinary}
\title{Random data for binary classification example.}
\format{A data.frame with 1,000 rows and 4 variables:
\itemize{
  \item x1: feature 1
  \item x2: feature 2
  \item x3: feature 3
  \item target: binary target variable
}}
\usage{
data("dfBinary")
}
\description{
Random data for binary classification example.
}
\examples{
data("dfBinary")
}
\keyword{datasets}
